refactor(engine-host): move session construction/resolution/persistence below the seam
Split the mixed rusty-sudocode-cli/src/cli/session.rs into its engine and
renderer halves. The engine half moves to engine_host::session: SessionHandle,
current_session_store, the new/create/resolve/load session helpers,
delete_managed_session (with its offloaded-tool-results GC), and
write_session_clear_backup. These wrap runtime::SessionStore -- the transcript
store the turn loop reads and writes -- so both the REPL and engine-acp build
and resolve a session through them.

The renderer half stays in cli::session: the session-list / picker / delete-
confirmation UI, plus list_managed_sessions / ManagedSessionSummary, which are
coupled to cli::lifecycle (SessionLifecycleSummary) and would form an
engine-host -> rusty-sudocode-cli cycle if moved. They import the relocated
current_session_store from engine_host::session (renderer -> engine-host, the
allowed direction). SessionHandle is a leaf {id, path}; the CORE AcpCliSession
holds it beside its BuiltRuntime, so it extracts cleanly before the core.

Also deletes three symbols with zero callers crate-wide (verified):
resolve_managed_session_path, latest_managed_session, SESSION_REFERENCE_ALIASES
-- rather than propagate dead code into the new crate. No new engine-host deps
(runtime only); the delete-GC unit test moves verbatim.

Third staged step of the engine-host extraction (#120).
