+++ rust/crates/rusty-sudocode-cli/src/cli/session.rs
@@ -1,20 +1,12 @@
-use std::fs;
 use std::io::{self, Write};
-use std::path::{Path, PathBuf};
+use std::path::PathBuf;
 use std::time::UNIX_EPOCH;
 
-use runtime::{Session, SessionStore};
+use engine_host::session::current_session_store;
 
 use crate::cli::lifecycle::{classify_session_lifecycle_for, SessionLifecycleSummary};
 
 pub(crate) const LATEST_SESSION_REFERENCE: &str = "latest";
-pub(crate) const SESSION_REFERENCE_ALIASES: &[&str] = &[LATEST_SESSION_REFERENCE, "last", "recent"];
-
-#[derive(Debug, Clone)]
-pub(crate) struct SessionHandle {
-    pub(crate) id: String,
-    pub(crate) path: PathBuf,
-}
 
 #[derive(Debug, Clone)]
 pub(crate) struct ManagedSessionSummary {
@@ -30,59 +22,6 @@ pub(crate) struct ManagedSessionSummary {
 
 pub(crate) fn sessions_dir() -> Result<PathBuf, Box<dyn std::error::Error>> {
     Ok(current_session_store()?.sessions_dir().to_path_buf())
-}
-
-pub(crate) fn current_session_store() -> Result<SessionStore, Box<dyn std::error::Error>> {
-    let cwd = runtime::current_workspace_root()?;
-    SessionStore::from_cwd(&cwd).map_err(|e| Box::new(e) as Box<dyn std::error::Error>)
-}
-
-pub(crate) fn new_cli_session() -> Result<Session, Box<dyn std::error::Error>> {
-    new_cli_session_for(&runtime::current_workspace_root()?)
-}
-
-pub(crate) fn new_cli_session_for(cwd: &Path) -> Result<Session, Box<dyn std::error::Error>> {
-    Ok(Session::new().with_workspace_root(cwd.to_path_buf()))
-}
-
-pub(crate) fn create_managed_session_handle(
-    session_id: &str,
-) -> Result<SessionHandle, Box<dyn std::error::Error>> {
-    let cwd = runtime::current_workspace_root()?;
-    create_managed_session_handle_for(&cwd, session_id)
-}
-
-pub(crate) fn create_managed_session_handle_for(
-    cwd: &Path,
-    session_id: &str,
-) -> Result<SessionHandle, Box<dyn std::error::Error>> {
-    let handle = SessionStore::from_cwd(cwd)
-        .map_err(|e| Box::new(e) as Box<dyn std::error::Error>)?
-        .create_handle(session_id);
-    Ok(SessionHandle {
-        id: handle.id,
-        path: handle.path,
-    })
-}
-
-pub(crate) fn resolve_session_reference(
-    reference: &str,
-) -> Result<SessionHandle, Box<dyn std::error::Error>> {
-    let handle = current_session_store()?
-        .resolve_reference(reference)
-        .map_err(|e| Box::new(e) as Box<dyn std::error::Error>)?;
-    Ok(SessionHandle {
-        id: handle.id,
-        path: handle.path,
-    })
-}
-
-pub(crate) fn resolve_managed_session_path(
-    session_id: &str,
-) -> Result<PathBuf, Box<dyn std::error::Error>> {
-    current_session_store()?
-        .resolve_managed_path(session_id)
-        .map_err(|e| Box::new(e) as Box<dyn std::error::Error>)
 }
 
 pub(crate) fn list_managed_sessions(
@@ -104,57 +43,6 @@ pub(crate) fn list_managed_sessions(
             lifecycle: lifecycle.clone(),
         })
         .collect())
-}
-
-pub(crate) fn latest_managed_session() -> Result<ManagedSessionSummary, Box<dyn std::error::Error>>
-{
-    let store = current_session_store()?;
-    let lifecycle = classify_session_lifecycle_for(store.workspace_root());
-    let session = store
-        .latest_session()
-        .map_err(|e| Box::new(e) as Box<dyn std::error::Error>)?;
-    Ok(ManagedSessionSummary {
-        id: session.id,
-        path: session.path,
-        updated_at_ms: session.updated_at_ms,
-        modified_epoch_millis: session.modified_epoch_millis,
-        message_count: session.message_count,
-        parent_session_id: session.parent_session_id,
-        branch_name: session.branch_name,
-        lifecycle,
-    })
-}
-
-pub(crate) fn load_session_reference(
-    reference: &str,
-) -> Result<(SessionHandle, Session), Box<dyn std::error::Error>> {
-    let loaded = current_session_store()?
-        .load_session(reference)
-        .map_err(|e| Box::new(e) as Box<dyn std::error::Error>)?;
-    Ok((
-        SessionHandle {
-            id: loaded.handle.id,
-            path: loaded.handle.path,
-        },
-        loaded.session,
-    ))
-}
-
-pub(crate) fn delete_managed_session(path: &Path) -> Result<(), Box<dyn std::error::Error>> {
-    if !path.exists() {
-        return Err(format!("session file does not exist: {}", path.display()).into());
-    }
-    fs::remove_file(path)?;
-    // GC the session's offloaded tool-results subtree (write-once blobs spilled
-    // from oversized tool outputs). The managed file is `<session-id>.jsonl`, so
-    // its stem is the session-id that namespaces the offload dir. Best-effort:
-    // an absent dir (session never offloaded anything) is not an error.
-    if let Some(session_id) = path.file_stem().and_then(|stem| stem.to_str()) {
-        if let Some(dir) = Session::tool_results_dir_for(path, session_id) {
-            let _ = fs::remove_dir_all(dir);
-        }
-    }
-    Ok(())
 }
 
 pub(crate) fn confirm_session_deletion(session_id: &str) -> bool {
@@ -257,60 +145,5 @@ pub(crate) fn format_session_modified_age(modified_epoch_millis: u128) -> String
         60..=3_599 => format!("{}m-ago", delta_seconds / 60),
         3_600..=86_399 => format!("{}h-ago", delta_seconds / 3_600),
         _ => format!("{}d-ago", delta_seconds / 86_400),
-    }
-}
-
-pub(crate) fn write_session_clear_backup(
-    session: &Session,
-    session_path: &Path,
-) -> Result<PathBuf, Box<dyn std::error::Error>> {
-    let backup_path = session_clear_backup_path(session_path);
-    session.save_to_path(&backup_path)?;
-    Ok(backup_path)
-}
-
-pub(crate) fn session_clear_backup_path(session_path: &Path) -> PathBuf {
-    let timestamp = std::time::SystemTime::now()
-        .duration_since(UNIX_EPOCH)
-        .ok()
-        .map_or(0, |duration| duration.as_millis());
-    let file_name = session_path
-        .file_name()
-        .and_then(|value| value.to_str())
-        .unwrap_or("session.jsonl");
-    session_path.with_file_name(format!("{file_name}.before-clear-{timestamp}.bak"))
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use std::time::{SystemTime, UNIX_EPOCH};
-
-    #[test]
-    fn delete_managed_session_gcs_offloaded_tool_results() {
-        let nanos = SystemTime::now()
-            .duration_since(UNIX_EPOCH)
-            .expect("time")
-            .as_nanos();
-        let sessions_dir = std::env::temp_dir().join(format!("scode-del-gc-{nanos}"));
-        fs::create_dir_all(&sessions_dir).expect("sessions dir");
-        let session_id = "sess-abc-123";
-        let session_file = sessions_dir.join(format!("{session_id}.jsonl"));
-        fs::write(&session_file, b"{}\n").expect("write session file");
-
-        // An offloaded blob under the session's namespaced tool-results dir.
-        let offload_dir =
-            Session::tool_results_dir_for(&session_file, session_id).expect("offload dir");
-        fs::create_dir_all(&offload_dir).expect("offload dir");
-        fs::write(offload_dir.join("tool-1"), b"big output").expect("write blob");
-        assert!(offload_dir.exists());
-
-        delete_managed_session(&session_file).expect("delete should succeed");
-
-        // Both the transcript and the offloaded subtree are gone.
-        assert!(!session_file.exists());
-        assert!(!offload_dir.exists());
-
-        let _ = fs::remove_dir_all(&sessions_dir);
     }
 }
